have twitter arr repeated
