+++ package.json
@@ -34,6 +34,7 @@
         "morgan": "^1.9.1",
         "node-fetch": "^2.3.0",
         "nodemon": "^1.18.9",
+        "serve-favicon": "^2.5.0",
         "twitter": "^1.7.1"
     }
 }
