+++ src/views/helpers/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+    getData: require('./getData')
+};
